Overhaul, document, and rename
