+++ src/main/java/io/github/daflamingfox/CommandMessage.java
@@ -1,4 +1,4 @@
-package me.hopedev.commandhandler;
+package io.github.daflamingfox;
 
 import java.util.ArrayList;
 import java.util.Arrays;
@@ -9,6 +9,12 @@ public class CommandMessage {
     private final String caller;
     private final String prefix;
 
+    /**
+     * Creates a CommandMessage object.
+     * splits the message into arguments
+     * @param commandMessage The message from the command.
+     * @param prefix The prefix from the command
+     */
     public CommandMessage(String commandMessage, String prefix) {
         this.prefix = prefix;
         commandMessage = commandMessage.replaceFirst(prefix, "");
@@ -17,18 +23,36 @@ public class CommandMessage {
         args.remove(0);
     }
 
+    /**
+     * 
+     * @return The caller, the prefix and invocator put together.
+     * @author Aurel Ballin
+     */
     public final String getCaller() {
         return this.caller;
     }
 
+    /**
+     * 
+     * @return The prefix.
+     */
     public final String getPrefix() {
         return this.prefix;
     }
 
+    /**
+     * 
+     * @param index The arg you want to get.
+     * @return The arg at index, index.
+     */
     public final String getArg(int index) {
         return this.args.get(index);
     }
 
+    /**
+     * 
+     * @return The ArrayList<String> of args.
+     */
     public final ArrayList<String> getArgs() {
         return this.args;
     }
