--- src/main/java/me/hopedev/commandhandler/CommandExecutor.java
@@ -1,9 +1,0 @@
-package me.hopedev.commandhandler;
-
-import java.util.ArrayList;
-
-public interface CommandExecutor {
-
-
-    void execute(final CommandData data, ArrayList<Command> commands);
-}
