--- src/main/java/me/hopedev/commandhandler/CommandPermissions.java
@@ -1,5 +1,0 @@
-package me.hopedev.commandhandler;
-
-public enum CommandPermissions {
-    // later use
-}
